+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - fixed cron logic to accurately cancel orders & mark orders shipped in Shopgate Panel
 - fixed product sort to use parent ID instead of child ID for sort order positions in a category
 - fixed product export not honoring clearance flag
+- fixed category sort cache is now more responsive to cache clears
 
 # 1.8.5